Enhance dataset configurations and model patching for improved functionality
- Added new dataset entries for 'identity_sean_dpo' and 'identity_sean_dpo_improved' in dataset_info.json, including formatting and ranking specifications.
- Updated patch_model function in patcher.py to ensure proper handling of model.generate when missing __func__.
- Implemented a minimal GGUFLoader in load_kt_peft_model to maintain device_map for inference when gguf_loader is not set.
- Enhanced Dockerfile for kt-sft with version control and improved installation steps for dependencies.
- Updated evaluation scripts to include comprehensive benchmarks for assessing model capabilities, including MMLU, HellaSwag, GSM8K, and TruthfulQA.
- Improved documentation for training and evaluation processes, ensuring clarity on new features and configurations.

+++ kt-sft/ktransformers/operators/experts.py
@@ -877,9 +877,10 @@ class KSFTExpertsCPU(torch.autograd.Function):
         if weights.device.type != 'cpu':
             if debug:
                 print(f"[KSFTExpertsCPU.backward] weights is on {weights.device}, moving to CPU", flush=True)
-            weights = weights.cpu().clone().contiguous()
+            weights = weights.to(torch.float32).cpu().clone().contiguous()
         else:
-            weights = weights.contiguous()
+            # Convert to float32 as C++ expects const float*
+            weights = weights.to(torch.float32).contiguous()
             
         # Create input_grad explicitly on CPU with explicit shape and dtype
         # Use pin_memory=True like the rest of the codebase to ensure C++ can access it
